+++ src/components/Button/Button.tsx
@@ -1,6 +1,5 @@
 import * as React from "react";
 import { fontSize } from "../Typography/Typography";
-import { Props as BaseButtonProps } from "./internal/BaseButton";
 import * as Styled from "./styled";
 import { Theme, useTheme } from "../../themes";
 import { getShadow } from "../../utils/getShadow";
@@ -166,8 +165,7 @@ const getPadding = ({ theme, size, color }: Padding) => ({
   paddingBottomAtActive:
     color === "clear" ? "" : paddingAtActive[size].paddingBottom,
 });
-
-export type ButtonProps = Omit<BaseButtonProps, "color"> & {
+type baseProps = {
   /**
    * The component used for the root node.
    * Default: `<button />`
@@ -182,11 +180,22 @@ export type ButtonProps = Omit<BaseButtonProps, "color"> & {
   inline?: boolean;
   size?: ButtonSize;
   onClick?: (event: React.MouseEvent<Element, MouseEvent>) => void;
+} & React.ComponentPropsWithoutRef<"button">;
+type AnchorProps = Omit<baseProps, "onClick"> & {
   /**
    * If added this props, root node becomes `<a />`.
    */
-  href?: string;
+  href: string;
+  target?: "_blank" | "_self" | "_parent" | "_top";
+  rel?: string;
 };
+type BaseButtonProps = baseProps & {
+  href?: undefined;
+  target?: undefined;
+  rel?: undefined;
+};
+
+export type ButtonProps = BaseButtonProps | AnchorProps;
 
 const Button = React.forwardRef<HTMLElement, ButtonProps>(function Button(
   {
@@ -195,7 +204,6 @@ const Button = React.forwardRef<HTMLElement, ButtonProps>(function Button(
     color = "primary",
     inline = false,
     size = "medium",
-    href,
     ...rest
   },
   ref,
@@ -213,14 +221,25 @@ const Button = React.forwardRef<HTMLElement, ButtonProps>(function Button(
     color,
   });
 
-  const isLink = !!href;
-  let anchorProps: any = {};
-  if (isLink) {
-    anchorProps = {
-      as: component || "a",
-      href,
-    };
-  }
+  const anchorProps: {
+    as?:
+      | keyof JSX.IntrinsicElements
+      | React.ComponentType<{ className: string }>;
+    href?: string;
+    target?: "_blank" | "_self" | "_parent" | "_top";
+    rel?: string;
+  } = (() => {
+    if (typeof rest.href === "string") {
+      const { href, target, rel } = rest;
+      return {
+        as: component || "a",
+        href,
+        target,
+        rel,
+      };
+    }
+    return {};
+  })();
 
   return (
     <Styled.ButtonContainer