+++ src/components/NavigationRail/Menu/styled.ts
@@ -6,10 +6,7 @@ export const Container = styled.div<{ isActive: boolean }>`
   cursor: pointer;
   display: flex;
   align-items: center;
-  padding: ${({ theme }) =>
-    `${theme.spacing * 2}px ${theme.spacing * 2}px ${theme.spacing * 2}px ${
-      theme.spacing * 3
-    }px`};
+  padding: ${({ theme }) => `${theme.spacing * 2}px`};
   background-color: ${({ isActive, theme }) =>
     isActive ? theme.palette.background.hint : "none"};